Fixed login/signup standards
