--- src/pages/Login.tsx
@@ -1,11 +1,0 @@
-import "./pages.css";
-
-const Login = () => {
-  return (
-    <div className="page-layout">
-      <h1>Login</h1>
-    </div>
-  );
-};
-
-export default Login;
