+++ src/utils/validations.tsx
@@ -1,0 +1,3 @@
+export const emailRegex = /^[\w.-]+@[\w.-]+\.[A-Za-z]{2,}$/;
+export const passwordSignUpRegex = /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/;
+export const passwordLogInRegex = /^.{8,}$/;
